Attach new conversation to existing channel (fd.o #26644)
When there was a channel with a conversation/switchboard attached to it,
the switchboard got closed by the server after a certain timeout, then
if the other client send a message to us it opened a new switchboard
that we ignored since there was already an existing channel for the
contact. This fixes it by attaching the incoming switchboard to the
existing channel.