conference: fix extending from 1-1 channels where the switchboard has timed out
Previously, if you set InitialChannels to [C1, C2], then the
switchboard of C1 would be stolen and the participants of C2 were
added to C1's switchboard.

However, if C2's switchboard had timed out, C2's participant list
would be empty. So, instead of using the Conversation.participants
set, define a new function that the 1-1 channel can override to be
sure it always returns the correct contact.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>