im/conference channel: raise NotAvailable when started with an invalid contact
If you create a text channel with TargetID="ajksdad" then it'll work
because butterfly just thinks the contact is offline. If there is no
contact available from looking in the addressbook, then we won't even
be able to send offline messages, so fail when creating the channel.

Ditto for the conference channel when specifying initial invitees.

Fixes fd.o#27122

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>