connection: only switch to HTTP mode if we failed during connection
If we're connected fine directly, and then the cable is pulled, we
don't want to try and connect immediately back using an HTTP
connection, otherwise calling logout will trigger an
AssertionError[0].

Additionally, we just don't want that to happen anyway because it's
annoying. If the cable is pulled we need a NetworkError to appear
instead.

0. http://launchpadlibrarian.net/44282269/butterfly.txt

Fixes fd.o#26147

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>