Be explicit about data types in methods that produce contact attributes
Previously, the wrong type was returned for SimplePresence/presence.
Until telepathy-glib 0.11.3, this was masked by the fact that TpContact
would respond to missing or wrongly-typed information in the result of
GetContactAttributes by falling back to a slower path. This is no longer
done, meaning that all Butterfly contacts appear to have UNKNOWN presence.

Reviewed-by: Olivier Le Thanh Duong <olivier@lethanh.be>