bugfix: correctly normalize and ensure handles
We use new methods of tp-python to ensure a handle exists (it's created if needed). That means only the connection keeps a set of all the handles and there is no more dict persistent between connections.

We make sure the handle name is normalized. This is particularly critical for contact handles because the account name might not always be in the same case.