Fix bug fd #37100.
The core problem is a crash in ButterflyTextChannel._signal_text_received.
The message gets in the pending messages queue, but no one acknowledges it,
plus the recv id doesn't increase because an exception is thrown.

The exceptions in my case were 'UnicodeDecodeError' and 'ValueError'.
The second exception was a result of the first one (because ID was
always 0), and the UnicodeDecodeError (that occured with message from
a received offline message) is a separate bugfix.